added two openneuro research studies
