+++ data-explore/studies/openneuro-ds004504/README.md
@@ -1,0 +1,9 @@
+# EEG Transform Pipeline (Research Use Only)
+
+**Purpose**:  
+Builds a reusable pipeline to transform raw EEG → PSD → residualized features → UMAP.  
+
+**⚠️ Disclaimer**:  
+- Groups ("Alzheimer’s", "FTD") reflect labels from [OpenNeuro ds006036](link).  
+- This tool is **not validated for clinical use**.  
+- By proceeding, you agree to use this only for research/hobby purposes.
